finishing preprocessor dev
