+++ preprocessors/mmsemseg/requirements.txt
@@ -1,4 +1,5 @@
 opencv-python==4.7.0.72
-jupyterlab==3.6.1
 flask==2.2.3
 jsonschema==3.2.0
+gunicorn==20.1.0
+jupyterlab
